Complete exercise 3
Did the exercise a completely different way than the way taken in the
solution. Their function takes a string, splits character by character
to an array, then it reverses the array and joins its elements to a new
string. My solution is simpler and less elegant. I just loop through the
string from its end and concatenate the string's elements to a new
string that's empty initially.